+++ digos-ambassador/Database/Appearances/Appearance.cs
@@ -42,7 +42,7 @@ namespace DIGOS.Ambassador.Database.Appearances
 	public class Appearance : IEFEntity
 	{
 		/// <inheritdoc />
-		public uint ID { get; set; }
+		public long ID { get; set; }
 
 		/// <summary>
 		/// Gets or sets the parts that compose this appearance.
